Move copyright and license as code comments
When generating the documentation, the License and Copyright text
appears on every module. It might be confusing for users if the license
applies to the code or the documentation.

This commit just moves the copyright notice and license specification as
a regular comment instead of putting it into the docstring that gets
rendered as documentation.

This was automatically adjusted using:

```sh
find -name '*.py' -not -path './.*' \
	| xargs sed -i '1 i\# License: MIT\n# Copyright © 2022 Frequenz Energy-as-a-Service GmbH\n'
find -name '*.py' -not -path './.*' \
	| xargs sed -zi 's/\n\?\(Copyright\|License\)\n[^\n]*\n\n\?//g'
```

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>