ci: Build and publish documentation
For pull requests the documentation is built and uploaded as an
artifact, so it can be manually checked.

For pushes, the documentation is built and published.

When pushing to a branch, if it is the default branch, then the
published version is called "next". At some point we might want to
consider having multiple vMAJOR.MINOR-dev documentation generated, but
will come when/if we have multiple vMAJOR.MINOR being maintained. For
now if a branch other than the default branch is pused to, nothing will
be published.

When pushing a new tag, if it is a pre-release tag (includes a `-`),
then the documentation is not published (again it could be published
once we have vMAJOR.MINOR-dev versions of the documentation). Otherwise
the documentation is published with version vMAJOR.MINOR and a vMAJOR
alias is added. If the MAJOR is the same as the default branch,
a "latest" alias is added/updated to that version too.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>