Run mypy on noxfile.py and be more restrictive
There is no reason to ignore imports, that only makes checking the types
weaker.

To be able to check types in a more strict way, we need to use packages
(-p) to avoid errors about having the same file in 2 packages
(src.frequenz.channels and frequenz.channels), so we need to translate
directory to package when running on `src`. Also we need to check
individual files separately.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>