Make receivers raise ReceiverError
The `Receiver.ready()` method (and related `receive()` and `__anext__`
when used as an async iterator) now `raise`s a `ReceiverError` and in
particular a `ReceiverStoppedError` when the receiver has no more
messages to receive.

`Receiver.consume()` doesn't raise any exceptions now (except for
`AssertionError` if there are programming errors).

Receivers raising `EOFError` now raise `ReceiverError` instead.

For channels which receivers stop receiving when the channel is closed,
the `ReceiverStoppedError` is chained with a `__cause__` that is a
`ChannelClosedError` with the channel that was closed.

Since now there is no need to use `ChannelClosedError` without
specifying a channel, now it requires the argument `channel` (before it
was optional).

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>