Make Select.ready() an async iterator
The async iterator yields a set of receivers that are ready to be
consumed. Users need to consume() explicitly from the receivers that
are ready and are not automatically consumed() by the select object if
they were not consumed in the select loop.

If a receiver is stopped, then it will be automatically removed from the
select loop.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>