Add method to stop Select
Add method to stop `Select` instance and cleanup any
pending task.
Select stores async tasks that listen on the receivers.
If select is removed without stopping the receivers tasks,
then we get error: Task was destroyed but it is pending!

We should stop all internal tasks before destoring Select.
Destructor can't be async, so we do that with extra method `stop`.
This method should be called when Select is no longer needed.

Signed-off-by: ela-kotulska-frequenz <elzbieta.kotulska@frequenz.com>