Add stop method for MergeNamed and Merge channels
Both channels stores async tasks that listen on the receivers.
If they are removed without stopping the receivers tasks,
then user gets error: Task was destroyed but it is pending!

`stop` method should be called when channel is no longer needed.
Destructor can't be async, so we do that with extra method `stop`.

Signed-off-by: ela-kotulska-frequenz <elzbieta.kotulska@frequenz.com>