Add a periodic timer receiver
The current timer provided is very convenient to implement timers, but
it's not very helpful to implement a peridic timer that doesn't
accumulate drift.

This commit adds a `util.PeriodicTimer` that ensures that drift is not
accumulated and that the receiver will be ready exactly after one
`interval` has passed (not counting any delay introduced by the event
loop being busy with another task).

It is also possible to control when this timer is started, and can be
stopped and reset/restarted.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>