Set the default tolerance to 0 for SKIP_AND_DRIFT
This is the most sensible default for this tick behaviour.

At some point we might want to refactor this so behaviours are classes
instead of an enum, so they can take their own options (with the
defaults that make more sense) and even calculate the next tick on their
own.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>