Convert MissedTickBehavior to classes
There are several advantages of using classes instead of enum for the
missed tick behavior:

* The `delay_tolerance` option doesn't make sense for all behavior, so
  now each behavior can have its own parameters
* New behaviors can be implemented without having to modify the
  `PeriodicTimer`
* Easier to test

This commit doesn't update the tests because when trying to update them
and use the `hypothesis` package to test the different behaviors, it was
evident that using floats to represent times is a recipe for disaster,
specially when the monotonic clock is way bigger than the interval. For
behaviors when more math than a simple addition is needed (for example,
modulo), it is very likely to have wrong calculations on those
circumstances.

It is even necessary to adapt the current tests to use a bit less than
the tolerance in some case to cope with floating point errors.

A following commit will use `int`s internally to represent the time,
intervals and drifts.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>