Use int to represent time internally
As pointed out in the previous commit, using float to represent time
can lead to extremely hard to debug issues in cases where the magnitude
of the amounts to compare or operate with, or the amount of relevant
digits are very different.

To avoid this we use ints to represent time internally, all in
microseconds.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>