Add `FileWatcher` unit tests
The existing tests are really integration tests, as they are not only
testing `FileWatcher`, but also `Select` and `PeriodicTimer`, which is
confusing when there is a bug in any of the other classes as we might
get test failures here.

This commit adds a a proper unit tests mocking `awatch`.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>