Fix FileWatcher skipping DELETE events
The FileWatcher was skipping DELETE events, except on very weird cases
where the file was deleted and re-created just before the event was
triggered by awatch().

This also extends the integration test to test both situations (delete
events when a file exist and when it doesn't).

Because there is usually a considerable delay between a change in the
filesystem happens and awatch() emits an event it is difficult to decide
if we should filter events based on the file existing or not. For
example, if a create event is received but before is triggered the file
was removed, should we emit the create event or not? Or the other way
around, should we emit a delete event if a file was created again before
the event was triggered.

It is probably best to leave this decision to the user, and let them
deal with races.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>