Report the type of change observed in FileWatcher
Previously, the FileWatcher only provided information of the
path of the changed file, without indicating whether the change
was a creation, modification, or deletion. This commit improves the
functionality of the FileWatcher by adding the ability to report
the type of change observed. This additional information can be
useful when tracking changes to files and understanding the
context of the change.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>