Update event_types parameter in FileWatcher
Previously, the event_types parameter in FileWatcher accepted
either a set or None value, with None used as a mechanism to
prevent issues related to Python mutable objects and default
values.

This commit simplifies the code by removing the handling of
None as a default value for event_types while still preventing
issues related to mutable objects.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>