Test docs generation more widely
We allow the docs generation test to run for any trigger but `push`, so
for exampmle `workflow_dispatch` will also run the test.  We rename the
job accordingly also to `test-generate-docs`.

The `push` event is still skipped because it already runs a job to
generate the final documentation, so we don't need to test it too.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>