Trigger CI workflow on merge queues
For this to work properly we need to exclude the branches starting with
`gh-readonly-queue/` for pushes, otherwise PRs will be tested twice,
once as the PR and the second one when they are pushed to the
`gh-readonly-queue/...` branch.

We do run the tests for the merge queue merge commit as it might include
several PRs being merged at the same time, so we want to test the merge.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>