Add a new async generator implementation for select()
It turns out the *broken* async generators are not completely leaked, it
just needs more "time" to be cleaned up.

In the example in 9015e7da0f9faeb2bae470d62c040f9ea1839ceb if a few
`await asyncio.sleep(0)` are added, then the generator is properly
finalized. Even when the destruction is much less deterministic than
using a context manager, it should be safe enough to assume that in a
long running program the generator will be eventually (and very soon)
cleaned up, and the user interface is much simpler and intuitive, so
we decided to bring it back.

This in a way reverts 9015e7da0f9faeb2bae470d62c040f9ea1839ceb but it
also simplifies the code and fixes a few bugs.

The tests of the selector are ported to test select() and we make sure
all tasks are properly cleaned up after a few yields to the event loop.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>