Add an optional start delay to `Timer`
`Timer()`, `Timer.timeout()`, `Timer.periodic()` and `Timer.reset()`
now take an optional `start_delay` option to make the timer start after
some delay.

This can be useful, for example, if the timer needs to be *aligned* to
a particular time. The alternative to this would be to `sleep()` for the
time needed to align the timer, but if the `sleep()` call gets delayed
because the event loop is busy, then a re-alignment is needed and this
could go on for a while. The only way to guarantee a certain alignment
(with a reasonable precision) is to delay the timer start.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>