Broadcast: Use `hash()` instead of a UUID
UUIDs are expensive to create and not really necessary in this case as
we are just keeping a local list of unique objects, so using the
object's hash is enough.

We also use a `id(self)`-based name by default if a `name` was not
provided when creating a `new_receiver()`.

A simple test using `timeit` shows a 2 orders of magnitude improvement
in Python 3.11:

```console
$ python -m timeit -s "import uuid" "uuid.uuid4()"
100000 loops, best of 5: 2.67 usec per loop
$ python -m timeit "hash(__name__)"
10000000 loops, best of 5: 19.6 nsec per loop
$ python -m timeit "id(__name__)"
10000000 loops, best of 5: 20.1 nsec per loop
```

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>