Add descriptive `str` and `repr` implementations
When debugging and logging objects it is very useful to get a
descriptive and clear string representation.

The new representation uses the class name and the user defined name
(if any) for `str` and a `repr` that tries to show how the class was
created but also important internal state.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>

# ------------------------ >8 ------------------------
# Do not modify or remove the line above.
# Everything below it will be ignored.
#
# Conflicts:
#	src/frequenz/channels/_base_classes.py
#	src/frequenz/channels/_bidirectional.py