Remove `Peekable` and associated methods/classes
`Peekable` was a class that allowed users to peek at the latest value in
a channel, without consuming anything. Even when useful in principle, it
is a special case that breaks the whole channel abstraction, and can
easily be emulated by just using a normal receiver and caching the last
value.

Also the following symbols are removed as they are not longer used:

* 'Broadcast.new_peekable`
* `Receiver.into_peekable`
* `ReceiverInvalidatedError`

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>