Improve string representation of merge receiver
Since `merge()` uses the `_Merge` class, when printing what's returned
by `merge()` the string representation of `_Merge` is used. This is not
very useful since it doesn't show the name that the user wrote in the
code.

This commit adds a `name` argument to `_Merge` and uses it to create the
string representation of the receiver. If no name is provided, the class
name is used as before.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>