Raise an exception is `merge()` is called with no receivers
Calling `merge()` with no receivers makes returns a stopped receiver,
which can be unexpected if it was done accidentally, as the user will
probably expect the merge receiver to block while receiving at least
once.

By raising an exception we make sure that users express their intentions
explicitly instead, having to write something like:

```python
receivers: list[Receiver[T]]
...
if receivers:
    async for m in merge(*receivers):
        ...
```

In this case it is clear the user is having the no-receivers case into
account and will handle it properly.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>