Make `Merge` public again
The `Merge` receiver was made private to try to hide the implementation
details to the user, but it was a bad idea because the `Merge` receiver
has more methods than a plain `Receiver`, in particular `stop()`, which
users need to use for guaranteed cleanup.

This also means now the `merge()` function returns the `Merge` instance
instead of an abstract `Receiver`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>