Add the `markdown-svgbob` extension
This extension allows documentation to include ASCII diagrams/graphs
that are rendered as SVG when the documentation is generated, giving a
nicer look to the diagrams/graphs while keeping the documents readable
in text form.

This is especially useful for docstrings, are they will be seen by users
in text form, but we still want to render them nicely.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>