Improve `Anycast` class documentation
Add a "Characteristics" summary with the most important properties of
this channel type, a diagram to show how messages are routed from
multiple senders to multiple receivers, explain a bit more about what
happens when buffers are full and improve and add a new example.

The formatting is done having in mind this documentation is now included
in the User Guide.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>