Improve the `_receiver` module description
Give a general introduction to receivers, explaining how one usually
create them, how to map values, how to handle receiving errors and how
to use the low-level methods.

This module is not publicly available, so users won't be able to access
the information in IDEs for example, but it will be rendered as part of
the User Guide.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>