Improve the `_select` module description
Explaining how `select()` should be used, including how to stop the
iteration, how to handle errors and mentioning the exhaustiveness check.

This module is not publicly available, so users won't be able to access
the information in IDEs for example, but it will be rendered as part of
the User Guide.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>