Improve the `timer` module documentation
Add a quick start guide with a couple of examples in the module
documentation, as well as a general explanation about missing ticks.

Improve documentation of the `MissedTickPolicy` class to explain how to
create custom policies, including an example. Also improve the
documentation of each missed tick policy provided by the module.

Improve the `Timer` class documentation to focus on how to use timers
more generally, referring to the `timeout()` and `periodic()`
constructors for more information, and improve both constructors
documentation to explain what they are useful for and include examples.

The formatting is done having in mind this documentation is now included
in the User Guide.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>