README: Make admonitions more readable in GitHub
Admonitions are not properly rendered by GitHub, so they don't look
great there, but they should be readable. Without leaving an empty line
they are rendered as a single paragraph, which is very messy.

Since the extra indentation is interpreted as a code block, we also need
to treat examples differently, as otherwise the markdown is rendered
verbatim, missing syntax highlighting and other features.

To overcome this issue we use simple headings for examples in the README
and then include them in the documentation as admnonitions. Becuase
of this we need to have more fine-grained include blocks in the README,
which is also not great, but it's better than the alternative.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>