Remove `Timer.periodic()` and `Timer.timeout()`
These names proved to be too confusing for users. It seems like having
a short name that can convey all the intricacies of timers in the async
world is very difficult, so we better stop trying to hide the complexity
and force users to read about and use missed tick policies explicitly.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>