Use GitHub Alert syntax in the README
This allows admonitions/callouts/alerts to be rendered properly both in
GitHub and `mkdocs`. On the downside, GitHub syntax is less flexible
than `mkdocs`, so we can't have the admonitions be rendered *inline end*
any more.

Because of this we only use the GitHub syntax for the README, as it is
more important that the README is rendered correctly in GitHub than
having it look a bit nicer in `mkdocs`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>