docs: Don't show boilerplate code in the use guide
We filter out all the boilerplate (importing, declaring some
variables) from the code shown in the user guide, as it is distracting
for such a high-level introduction. This means more examples listed in
modules have the boilerplate removed from the generated docs.

Examples in classes or functions, that usually show up in the API docs
are kept as before, with all the code. This is so users wanting to dive
deeper into the library can still copy&paste code snippets from these
examples and expect them to mostly work.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>