Remove `SelectErrorGroup`
This exception is very niche, is not really intended to be caught by
the users, as it should only happen when there is something really
wrong.

If there is any unexpected error during finalization, we just raise a
`BaseExceptionGroup`, which will be automatically converted to a regular
`ExceptionGroup` by Python if there are no `BaseException`s in the
group.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>