Move documentation from the module to the class
The module documentation is not being included in the rendered docs.
Normally this is done by including it in the User Guide, but since this
is an experimental module, we don't want to include it yet, so it is
easier to just move the docs to the class so they are rendered in the
API documentation.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>